pyproject.toml
[tool.poetry]
name = "fast-api"
version = "0.1.0"
description = ""
authors = ["k.melnikov <k.melnikov@foodtech.team>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.135.3"
uvicorn = "^0.43.0"
pydantic-settings = "^2.13.1"
aioboto3 = "^15.5.0"
pyjwt = {extras = ["crypto"], version = "^2.12.1"}

[tool.poetry.group.dev.dependencies]
boto3-stubs = ">=1.42.83,<1.43.0"
black = "^26.3.1"
flake8 = "^7.3.0"
pylint = "^4.0.5"
flake8-pyproject = "^1.2.4"
isort = "^8.0.1"
mypy = "^1.20.0"
pytest = "^9.0.2"
pytest-asyncio = "^1.3.0"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
max-line-length = 120
extend-ignore = ["E203", "F811"]

[tool.black]
line-length = 120
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 120

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
strict = true


[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
